Use native font stack (#417)
It ensures that the text is displayed in the font that the user is
most comfortable with, and it also reduces the size of the data
transferred.

Breaking change:

The Titillium web font has been replaced with native font stack.

Migration:

To keep the Titillium web font in your project, load the web font
and override the `--rui-font-family-base` custom property:

```
--rui-font-family-base: 'Titillium Web', helvetica, roboto, arial, sans-serif;
```