Reset the internal input value when the reset button is clicked in FileInputField (#641)
Not resetting the internal input value will cause the input to not
register any files if the user had clicked "Browse", then resetted the
field and again clicked "Browse" and selected a file. The 2nd file would
not have a event fired for this. This did not affect drag and drop.